Added class-based question demo
This greatly optimizes my previous code by outsourcing all the heavy lifting to a single class which I can call for each question, passing in the question, answers, and correct answer. The appearance does change a little bit, because I was lazy and decided to just print the array instead of formatting it properly, but it gets the job done; this isn't supposed to look pretty anyway.